Fix: "수정 토큰 claim 반환"
tokenService에서 토큰이 long 타입으로 반환하지 못하는 것을 수정하였습니다. 🚑

+++ demo/src/main/java/com/kh/demo/service/TokenService.java
@@ -67,7 +67,7 @@ public class TokenService {
 									.parse(token)
 									.getPayload();
 		
-		return (long) claims.get("memberNo");
+		return ((Integer) claims.get("memberNo")).longValue();
 	}
 	
 	public long parseBearerToken(String bearerToken) {
